pyproject.toml
[project]
name = "agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.77.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "langgraph-cli[inmem]>=0.4.12",
    "playwright>=1.57.0",
    "rich>=14.3.1",
]
